head: move stack out of .text
Stack area doesn't need to be measured by SKINIT. Initial %esp always
points to the end of SLB, there is no reason to move it from there.
Minimal stack size is checked by the code as the amount of available
space depends on size of SLRT passed by the bootloader.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>