event_log: obtain event log location from SLRT and log SKL measurements
SKL hashes are no longer passed from the bootloader. Measured code and
read-only data are not changed, and all other data is created at runtime
so SKL is able to measure itself, which it now does. Those hashes are
only used for event log entries, PCR was extended as a result of SKINIT
instruction on Dynamic Launch Event.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>