AMDSL: Add SKINIT reserved section and ensure SKL is 64KB long
A section of size 0x740 must the carved out just after the measured
portion of the SKL for the signature header, signature and public key.
The .fill directive ensures we have enough space in there and the
signing process does not crash into other sections.

In addition, adding ". = 0xFFF4" is a preliminary approach to ensure that
SKL is 64KB long. The extra missing 0xC bytes are produced by the .got.plt
section. A cleaner approach might be to use ALIGN() linker directives.

Signed-off-by: Jagannathan Raman <jag.raman@oracle.com>